requirements.txt
Flask==3.0.0
numpy==1.24.3
sounddevice==0.4.6
soundfile==0.12.1
webrtcvad==2.0.10
requests==2.31.0
edge-tts==6.1.9
playsound==1.3.0
faster-whisper==0.9.0
scikit-learn==1.3.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-dotenv==1.0.0 